requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
jinja2==3.1.4
hl7==0.4.5
fhir.resources==7.1.0
pydantic==2.10.3
pydantic-settings==2.6.1
lxml==5.3.0
openpyxl==3.1.5
xlrd==2.0.1
python-docx==1.1.2
aiofiles==24.1.0
httpx==0.27.2
python-jose==3.3.0
dicttoxml==1.7.16
xmltodict==0.14.2
